requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

numpy==1.26.4
decord==0.6.0
opencv-python-headless==4.10.0.84

torch
transformers>=4.49.0
accelerate>=0.33.0
bitsandbytes

openpack-toolkit; python_version < "3.12"
tqdm==4.66.4
pillow>=10.0.0